Move all types to the models directory
