+++ PHP/Collections/ReadOnlyCollection.php
@@ -1,7 +1,7 @@
 <?php
 namespace PHP\Collections;
 
-use PHP\Types\Type;
+use PHP\Types\Models\Type;
 
 /**
  * Defines an iterable set of read-only, key-value pairs
