+++ PHP/Types/Models/CallableType.php
@@ -1,5 +1,7 @@
 <?php
-namespace PHP\Types;
+namespace PHP\Types\Models;
+
+use PHP\Types\TypeNames;
 
 /**
  * Defines the base type for an item that can be invoked like a function
