+++ tests/TypesTest.php
@@ -22,9 +22,9 @@ class TypesTest extends TestCase
     public function testGetByNameReturnsType()
     {
         $this->assertInstanceOf(
-            'PHP\\Types\\Type',
+            'PHP\\Types\\Models\\Type',
             Types::GetByName( 'foobar' ),
-            'Expected Types::GetByName() to return a PHP\\Types\\Type instance'
+            'Expected Types::GetByName() to return a PHP\\Types\\Models\\Type instance'
         );
     }
     
@@ -78,9 +78,9 @@ class TypesTest extends TestCase
     public function testGetByValueReturnsType()
     {
         $this->assertInstanceOf(
-            'PHP\\Types\\Type',
+            'PHP\\Types\\Models\\Type',
             Types::GetByValue( null ),
-            'Expected Types::GetByValue() to return a PHP\\Types\\Type instance'
+            'Expected Types::GetByValue() to return a PHP\\Types\\Models\\Type instance'
         );
     }
     
@@ -115,9 +115,9 @@ class TypesTest extends TestCase
     public function testGetUnknownReturnsType()
     {
         $this->assertInstanceOf(
-            'PHP\\Types\\Type',
+            'PHP\\Types\\Models\\Type',
             Types::GetUnknownType(),
-            'Expected Types::GetUnknownType() to return a PHP\\Types\\Type instance'
+            'Expected Types::GetUnknownType() to return a PHP\\Types\\Models\\Type instance'
         );
     }
